Improve strategy validation and fix a calculation error in ARES V1
Add a new script for validating ARES V1 and V2 strategies, and correct a data type conversion issue in the ARES V1 strategy calculation.

+++ omnix_core/strategies/ares_v1.py
@@ -166,7 +166,7 @@ class AresQuantumProtocol:
                 kurtosis = np.mean((returns - np.mean(returns))**4) / (np.var(returns)**2)
                 quantum_vol = std_vol * (1 + 0.1 * abs(kurtosis - 3))
                 
-                return min(quantum_vol * 100, 100.0)  # Normalizar a 0-100%
+                return float(min(quantum_vol * 100, 100.0))  # Normalizar a 0-100%
             
             # Fallback: usar volatilidad simple
             return market_data.get('volatility', 50.0)
